Add container for main-right image
It seems the img tag is not a good container, so it was placed within the main-right container. Before this change, the img container didn't take the necessary space of the main container.

Now with the main-right container the image takes the necessary space. Not sure if there is any other way to fix this problem without the image being within a different container.