Restructure trigger container for better aligment
To be able to align the quoter with the quote text, a new container (the quote container) needed to be created so the quoter could be align at the end of the whole quote container. Before the change the quote and quoter text were in the container covering horizontally the whole page, but it was still necessary to apply the background color. I added a new container only for the size of the quoter and quote text so I could align the text the way I did and not lose the option of getting the background color applied horizontally taking all space available.